took `calculateMaxTextWidth` out to file
